refactor(p2p): extract engine tests
